Code affected: FileManager
Flaw: want to be able to pass in an abstract file tree and not build one from the existing filesystem (by passing in a string). This is mainly for ease of testing, though it might be useful in the future.

Fix: Added a constructor for FileManager that takes in a File, instead of a String.